requirements.txt
pywin32
flask
PyYAML